Change configs to use Doubles
Fixes config error in logs, need to double check 1.20.2

+++ src/armory/java/com/enderio/armory/common/config/common/ArmoryCommonConfig.java
@@ -12,16 +12,16 @@ public class ArmoryCommonConfig {
     public final ForgeConfigSpec.ConfigValue<Integer> EMPOWERED_EFFICIENCY_BOOST;
     public final ForgeConfigSpec.ConfigValue<Integer> EMPOWERED_ENERGY_PER_DAMAGE;
     public final ForgeConfigSpec.ConfigValue<Integer> EMPOWERED_ACTIVATION_COST_I;
-    public final ForgeConfigSpec.ConfigValue<Float> EMPOWERED_DAMAGE_ABSORPTION_CHANCE_I;
+    public final ForgeConfigSpec.ConfigValue<Double> EMPOWERED_DAMAGE_ABSORPTION_CHANCE_I;
     public final ForgeConfigSpec.ConfigValue<Integer> EMPOWERED_MAX_ENERGY_I;
     public final ForgeConfigSpec.ConfigValue<Integer> EMPOWERED_ACTIVATION_COST_II;
-    public final ForgeConfigSpec.ConfigValue<Float> EMPOWERED_DAMAGE_ABSORPTION_CHANCE_II;
+    public final ForgeConfigSpec.ConfigValue<Double> EMPOWERED_DAMAGE_ABSORPTION_CHANCE_II;
     public final ForgeConfigSpec.ConfigValue<Integer> EMPOWERED_MAX_ENERGY_II;
     public final ForgeConfigSpec.ConfigValue<Integer> EMPOWERED_ACTIVATION_COST_III;
-    public final ForgeConfigSpec.ConfigValue<Float> EMPOWERED_DAMAGE_ABSORPTION_CHANCE_III;
+    public final ForgeConfigSpec.ConfigValue<Double> EMPOWERED_DAMAGE_ABSORPTION_CHANCE_III;
     public final ForgeConfigSpec.ConfigValue<Integer> EMPOWERED_MAX_ENERGY_III;
     public final ForgeConfigSpec.ConfigValue<Integer> EMPOWERED_ACTIVATION_COST_IV;
-    public final ForgeConfigSpec.ConfigValue<Float> EMPOWERED_DAMAGE_ABSORPTION_CHANCE_IV;
+    public final ForgeConfigSpec.ConfigValue<Double> EMPOWERED_DAMAGE_ABSORPTION_CHANCE_IV;
     public final ForgeConfigSpec.ConfigValue<Integer> EMPOWERED_MAX_ENERGY_IV;
 
     public final ForgeConfigSpec.ConfigValue<Integer> EXPLOSIVE_RADIUS_ACTIVATION_COST_I;
@@ -56,16 +56,16 @@ public class ArmoryCommonConfig {
         EMPOWERED_EFFICIENCY_BOOST = builder.define("efficiencyBoost", 2);
         EMPOWERED_ENERGY_PER_DAMAGE = builder.define("energyUsePerDamagePoint", 750);
         EMPOWERED_ACTIVATION_COST_I = builder.define("activationCost_l1", 4);
-        EMPOWERED_DAMAGE_ABSORPTION_CHANCE_I = builder.define("damageAbsorptionChance_l1", 0.5f);
+        EMPOWERED_DAMAGE_ABSORPTION_CHANCE_I = builder.define("damageAbsorptionChance_l1", 0.5d);
         EMPOWERED_MAX_ENERGY_I = builder.define("maxEnergy_l1", 100000);
         EMPOWERED_ACTIVATION_COST_II = builder.define("activationCost_l2", 8);
-        EMPOWERED_DAMAGE_ABSORPTION_CHANCE_II = builder.define("damageAbsorptionChance_l2", 0.6f);
+        EMPOWERED_DAMAGE_ABSORPTION_CHANCE_II = builder.define("damageAbsorptionChance_l2", 0.6d);
         EMPOWERED_MAX_ENERGY_II = builder.define("maxEnergy_l2", 150000);
         EMPOWERED_ACTIVATION_COST_III = builder.define("activationCost_l3", 12);
-        EMPOWERED_DAMAGE_ABSORPTION_CHANCE_III = builder.define("damageAbsorptionChance_l3", 0.7f);
+        EMPOWERED_DAMAGE_ABSORPTION_CHANCE_III = builder.define("damageAbsorptionChance_l3", 0.7d);
         EMPOWERED_MAX_ENERGY_III = builder.define("maxEnergy_l3", 250000);
         EMPOWERED_ACTIVATION_COST_IV = builder.define("activationCost_l4", 16);
-        EMPOWERED_DAMAGE_ABSORPTION_CHANCE_IV = builder.define("damageAbsorptionChance_l4", 0.85f);
+        EMPOWERED_DAMAGE_ABSORPTION_CHANCE_IV = builder.define("damageAbsorptionChance_l4", 0.85d);
         EMPOWERED_MAX_ENERGY_IV = builder.define("maxEnergy_l4", 1000000);
         builder.pop();
